django-social-media-rest-api-master/app/requirements.txt
coverage==4.3.4
     django==2.2.3
     djangorestframework==3.7.7
 coreapi==2.3.3
 django-cors-headers
     drf_writable_nested==0.5.1
    
 djangorestframework_simplejwt==4.3.0